+++ NEWS.md
@@ -1,8 +1,20 @@
-## rlistings 0.2.5.9002
-* Fixed bug in pagination preventing key column values to appear in paginated listings when `export_as_txt` was used.
-* Added tests to cover for `export_as_txt` outputs.
+## rlistings 0.2.7.9008
+ * Added relevant tests for pagination when key columns need to be repeated in each page and when they are all empty.
+ * Added relevant tests for new line characters' handling in footnotes and titles.
 * Added function `split_listing_by_var` to split a listing into a list of listings according to values of a given 
   variable. This enables page splits by variable when paginating.
+
+## rlistings 0.2.7
+ * Applied `styler` and resolved package lint. Changed default indentation from 4 spaces to 2.
+ * Fixed bug in `add_listing_col` when both a function and a format are specified.
+ * Added a vignette on referential footnotes workaround.
+ * Added a vignette on formatting columns. 
+ * Added a vignette on pagination. 
+
+## rlistings 0.2.6
+ * Fixed bug in pagination preventing key column values to appear in paginated listings when `export_as_txt` was used.
+ * Added tests to cover for `export_as_txt` outputs.
+ * Integrated support for newline characters.
 
 ## rlistings 0.2.5
  * Fixed bug in `as_listing` preventing custom formatting from being applied to key columns.